# rest_service/api/swagger/swagger.yaml
openapi: 3.0.0
info:
  version: 3.0.0
  title: FFAStrans REST API Version 3.0
servers:
  - url: /
  - url: /api/json/v2
paths:
...
  /review:
    x-swagger-router-controller: review
    get:
      description: Gets review queue
      operationId: review
      responses:
        "200":
          description: Success
        default:
          description: Error
    delete:
      description: Deletes tickets from review queue
      parameters:
        - name: filename
          in: query
          description: File in review queue folder 
          required: true
          schema:
            type: string
      operationId: review_delete
      responses:
        "200":
          description: Success
        default:
          description: Error
  /tickets:
    x-swagger-router-controller: tickets
    get:
      description: Gets active and pending jobs
      parameters:
        - name: nodetails
          in: query
          description: Reduces performance impact on database significally, only returns currently active jobid's. Excluded incoming (those are complex to get)
          required: false
          schema:
            type: boolean
      operationId: tickets
      responses:
        "200":
          description: List of Jobs
        default:
          description: Error
  /machines:
    x-swagger-router-controller: machines
    get:
      description: Gets List of Farm Machines
      operationId: machines
      responses:
        "200":
          description: Success
        default:
          description: Error
    post:
      description: "Save about section"
      operationId: machines_post
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example: {"ip_address": ""}
      responses:
        "200":
          description: List of Machines and config
        default:
          description: Error   
    delete:
      description: Deletes machine by name. 
      parameters:
        - name: name
          in: query
          description: Machine name to delete 
          required: true
          schema:
            type: string
      operationId: machines_delete
      responses:
        "200":
          description: Success
        default:
          description: Error
...
  /getmediainfo:
    x-swagger-router-controller: get_mediainfo
    get:
      description: Get mediainfo about a file 
      operationId: get_mediainfo
      parameters:
        - name: filepath
          in: query
          description: The path to a file from perspective of the API server
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Success
        default:
          description: Error
  /workflows:
    x-swagger-router-controller: workflows
    get:
      description: ""
      operationId: workflows
      parameters:
        - name: id
          in: query
          description: Workflow ID, if not  specified returns all
          # problem of the old ffastrans api: parameters as part of url (not get params but url) must be required but ffastrans has optional params
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Success
        default:
          description: Error
    post:
      description: "Import Workflow"
      operationId: workflow_post
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: integer
                name:
                  type: string
              example: {"workflows": []}
      responses:
        "200":
          description: Success
        default:
          description: Error
  /variables:
    x-swagger-router-controller: variables
    get:
      description: ""
      operationId: variables
      responses:
        "200":
          description: Success
        default:
          description: Error
    delete:
      description: Deletes a variable
      operationId: variables_delete
      parameters:
        - name: name
          in: query
          description: Name of the variable to be deleted
          required: true
          schema:
            type: string
      
      responses:
        "200":
          description: Success
        default:
          description: Error
    post:
      description: "Create/Update Variable"
      operationId: variables_post
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: integer
                name:
                  type: string
              example: {"name":"s_example","data":"Example Description/Value"}
      responses:
        "200":
          description: Success
        default:
          description: Error
...